requirements.txt
python-Levenshtein
sqlalchemy
flask
flask-sqlalchemy 